requirements.txt
asgiref==3.7.2
cffi==1.16.0
crispy-bootstrap5==0.7
cryptography==41.0.7
Django==5.0
django-axes==6.3.0
django-crispy-forms==2.0
django-ratelimit==4.1.0
django-recaptcha==3.0.0
fuzzywuzzy==0.18.0
pycparser==2.21
python-dateutil==2.8.2
PyYAML==6.0.1
redis==5.0.1
setuptools==69.0.3
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
